Use correct format for NATURAL_32 and INTEGER_32.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@47805 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/C/run-time/out.c
@@ -642,7 +642,7 @@ rt_public EIF_REFERENCE c_outu(EIF_NATURAL_32 i)
 {
 	RT_GET_CONTEXT
 	register int len;
-	len = sprintf(buffero, "%lu", i);
+	len = sprintf(buffero, "%u", i);
 	return makestr(buffero, len);
 }
 
@@ -658,7 +658,7 @@ rt_public EIF_REFERENCE c_outi(EIF_INTEGER i)
 {
 	RT_GET_CONTEXT
 	register int len;
-	len = sprintf(buffero, "%ld", (long) i);
+	len = sprintf(buffero, "%d", i);
 	return makestr(buffero, len);
 }
